+++ include/effects/FadeInEffect.h
@@ -14,7 +14,7 @@ public:
             SmoothingFunction smoothingFunction = SmoothingFunction::SMOOTH_LINEAR
     );
 
-    void calculate(int startIndex, TemporaryLedData &tempData) override;
+    void calculate(float startIndex, TemporaryLedData &tempData) override;
 
     void reset() override;
 };
@@ -30,7 +30,7 @@ public:
             SamplingFunction samplingFunction = SamplingFunction::UNIFORM
             );
 
-    void calculate(int startIndex, TemporaryLedData &tempData) override;
+    void calculate(float startIndex, TemporaryLedData &tempData) override;
 
     void reset() override;
 };
