+++ include/effects/SolidEffect.h
@@ -11,19 +11,19 @@ protected:
 public:
     explicit SolidEffect(CRGB color, uint8_t opacity = 255);
 
-    void calculate(int startIndex, TemporaryLedData &tempData) override;
+    void calculate(float startIndex, TemporaryLedData &tempData) override;
 };
 
 
 class SolidSegmentEffect : public SolidEffect {
 
 protected:
-    int segmentLength = 0;
+    float segmentLength = 0;
 
 public:
-    explicit SolidSegmentEffect(CRGB color, int segmentLength, uint8_t opacity = 255);
+    explicit SolidSegmentEffect(CRGB color, float segmentLength, uint8_t opacity = 255);
 
-    void calculate(int startIndex, TemporaryLedData &tempData) override;
+    void calculate(float startIndex, TemporaryLedData &tempData) override;
 };
 
 }
