+++ src/effects/AbsolutePositionEffect.cpp
@@ -6,7 +6,7 @@ using namespace ledpipelines::effects;
 AbsolutePositionEffect::AbsolutePositionEffect(BaseLedPipelineStage *stage, int startPosition)
         : WrapperEffect(stage), startPosition(startPosition) {}
 
-void AbsolutePositionEffect::calculate(int startIndex, TemporaryLedData &tempData) {
+void AbsolutePositionEffect::calculate(float startIndex, TemporaryLedData &tempData) {
     this->stage->calculate(startPosition, tempData);
     this->state = this->stage->state;
 }
